requirements.txt
boto3==1.24.2
Django==4.1.1
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-storages==1.12.3
djangorestframework==3.13.1
docopt==0.6.2
google-api-python-client==2.43.0
google-auth-oauthlib==0.4.2
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.4
PyJWT==2.5.0
python-dotenv==0.21.0
python-environ==0.4.54
reportlab==3.6.12
drf-spectacular==0.12